pre comp (include error)

+++ push_swap_2/srcs/counter.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   counter.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hkoizumi <hkoizumi@student.42.jp>          +#+  +:+       +#+        */
+/*   By: xxxx <xxxx@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/06 12:22:39 by hkoizumi          #+#    #+#             */
-/*   Updated: 2024/03/21 13:29:45 by hkoizumi         ###   ########.fr       */
+/*   Updated: 2024/04/04 14:46:48 by xxxx             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/push_swap.h"
 
-int	node_counter(t_node *stack)
+long	node_counter(t_node *stack)
 {
-	int	node_count;
+	long	node_count;
 
 	if (!stack)
 		return (-1);
